Refactor: Split _write_only.py into separate files.
